+++ cmd/moombox/main.go
@@ -45,7 +45,7 @@ import (
 )
 
 var (
-	version = "2.5.1"
+	version = "2.5.2"
 	commit  = ""
 )